@@ -25,14 +25,17 @@
     "babel-runtime": "^5.8.20",
     "express": "^4.13.3",
     "js-yaml": "^3.4.0",
-    "react": "^0.14.3",
+    "less": "^2.7.1",
+    "react": "15.3.1",
+    "react-dom": "15.3.1",
     "sane": "^1.2.0",
     "webpack": "^1.12.1"
   },
   "dependencies": {
     "codemirror": "^5.6.0",
-    "codemirror-graphql": "^0.1.1",
-    "graphql": "^0.4.3",
+    "codemirror-graphql": "^0.5.7",
+    "graphql": "^0.7.0",
+    "graphql-tools": "^0.6.6",
     "marked": "^0.3.5"
   }
 }
