@@ -6,7 +6,11 @@
 
 var React = require('react');
 var Prism = require('./Prism');
-var Header = require('./Header');
+import Header from './Header';
+
+export default function Marked(props) {
+  return <div>{marked(props.children, props)}</div>;
+}
 
 /**
  * Block-Level Grammar
@@ -818,7 +822,29 @@ Parser.prototype.tok = function() {
       );
     }
     case 'code': {
-      return <Prism line={this.token.line}>{this.token.text}</Prism>;
+      if (this.token.lang === 'graphql') {
+        var lines = this.token.text.split('\n');
+        var firstLine = lines.shift().match(/^\s*#\s*({.*})$/);
+        if (firstLine) {
+          var metaData;
+          try {
+            metaData = JSON.parse(firstLine[1]);
+          } catch (e) {
+            console.error('Invalid Metadata JSON:', firstLine[1]);
+          }
+          if (metaData) {
+            var query = lines.join('\n');
+            var variables = metaData.variables ? JSON.stringify(metaData.variables, null, 2) : '';
+            return <script data-inline dangerouslySetInnerHTML={{__html: `
+              import MiniGraphiQL from '../_core/MiniGraphiQL';
+              import { StarWarsSchema } from '../_core/swapiSchema';
+              renderHere(<MiniGraphiQL schema={StarWarsSchema}
+              query={\`${query}\`} variables={\`${variables}\`} />);
+            `}} />
+          }
+        }
+      }
+      return <Prism language={this.token.lang} line={this.token.line}>{this.token.text}</Prism>;
     }
     case 'table': {
       var table = []
@@ -1084,11 +1110,3 @@ marked.InlineLexer = InlineLexer;
 marked.inlineLexer = InlineLexer.output;
 
 marked.parse = marked;
-
-var Marked = React.createClass({
-  render: function() {
-    return <div>{marked(this.props.children, this.props)}</div>;
-  }
-});
-
-module.exports = Marked;
