@@ -8,48 +8,27 @@
 
 var React = require('react');
 
+const links = [
+  { section: 'learn', text: 'Learn', href: '/learn/' },
+  { section: 'code', text: 'Code', href: '/code/' },
+  { section: 'community', text: 'Community', href: '/community/' },
+  { section: 'blog', text: 'Blog', href: '/blog/' },
+  { section: 'spec', text: 'Spec', href: 'http://facebook.github.io/graphql/' },
+];
+
 var HeaderLinks = React.createClass({
-  linksInternal: [
-    {section: 'docs', text: 'Docs', href: '/docs/getting-started/' },
-    {section: 'blog', text: 'Blog', href: '/blog/' },
-    {section: 'help', text: 'Help', href: '/help/' },
-    {section: 'code', text: 'Code', href: '/code/' },
-  ],
-  linksExternal: [
-    {section: 'spec', text: 'Spec', href: 'http://facebook.github.io/graphql/' },
-  ],
-
-  renderLinks: function(links) {
-    return links.map(function(link) {
-      return (
-        <li key={link.section}>
-          <a
-            href={link.href}
-            target={link.href.slice(0, 4) === 'http' ? '_blank' : null}
-            className={link.section === this.props.section ? 'active' : ''}>
-            {link.text}
-          </a>
-        </li>
-      );
-    }, this);
-  },
-
   render: function() {
-    return (
-      <div className="nav-site-wrapper">
-        <ul className="nav-site nav-site-internal">
-          {this.renderLinks(this.linksInternal)}
-        </ul>
-
-        <div className="algolia-search-wrapper">
-          <input id="algolia-doc-search" type="text" placeholder="Search docs..." />
-        </div>
-
-        <ul className="nav-site nav-site-external">
-          {this.renderLinks(this.linksExternal)}
-        </ul>
-      </div>
-    );
+    return <nav>
+      {links.map(link =>
+        <a
+          key={link.section}
+          href={link.href}
+          target={link.href.slice(0, 4) === 'http' ? '_blank' : null}
+          className={link.section === this.props.section ? 'active' : null}>
+          {link.text}
+        </a>
+      )}
+    </nav>;
   }
 });
 
